upgrade: update from cryptogarageinc/v0.3.16

+++ tools/cmake_cleanup.bat
@@ -5,13 +5,12 @@ if exist "cmake_cleanup.bat" (
   cd ..
 )
 
-rmdir /S /Q build
+if exist build rmdir /S /Q build
 
-rmdir /S /Q external\cfd
+if exist external\cfd rmdir /S /Q external\cfd
 
-rmdir /S /Q external\cfd-core
+if exist external\cfd-core rmdir /S /Q external\cfd-core
 
-rmdir /S /Q external\libwally-core
+if exist external\libwally-core rmdir /S /Q external\libwally-core
 
-rmdir /S /Q external\googletest
-
+if exist external\googletest rmdir /S /Q external\googletest
